+++ src/public/BaseFunctions.ps1
@@ -1,18 +1,17 @@
 # Mandatory function for the PackageManagement providers. It returns the name of your provider.
 function Get-PackageProviderName {
-	return $script:ProviderName
+	Write-Debug ($LocalizedData.ProviderDebugMessage -f ('Get-PackageProviderName'))
+	return 'WinGet'
 }
 
 # Mandatory function for the PackageManagement providers. It initializes your provider before performing any actions.
 function Initialize-Provider {
 	Write-Debug ($LocalizedData.ProviderDebugMessage -f ('Initialize-Provider'))
-
-	Update-PackageSources
-
 }
 
 # Defines PowerShell dynamic parameters so that a user can pass in parameters via OneGet to the provider
 function Get-DynamicOptions {
+	[Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSUseSingularNouns', '', Justification='Required by PackageManagement')]
 	param (
 		[Microsoft.PackageManagement.MetaProvider.PowerShell.OptionCategory]
 		$category
@@ -22,11 +21,11 @@ function Get-DynamicOptions {
 
 	switch ($category) {
 		Package {
-			Write-Output -InputObject (New-DynamicOption -Category $category -Name $script:additionalArguments -ExpectedType String -IsRequired $false)
+			Write-Output -InputObject (New-DynamicOption -Category $category -Name $script:AdditionalArguments -ExpectedType String -IsRequired $false)
 			Write-Output -InputObject (New-DynamicOption -Category $category -Name $script:AcceptLicense -ExpectedType Switch -IsRequired $false)
 		}
 		Install {
-			Write-Output -InputObject (New-DynamicOption -Category $category -Name $script:additionalArguments -ExpectedType String -IsRequired $false)
+			Write-Output -InputObject (New-DynamicOption -Category $category -Name $script:AdditionalArguments -ExpectedType String -IsRequired $false)
 			Write-Output -InputObject (New-DynamicOption -Category $category -Name $script:AcceptLicense -ExpectedType Switch -IsRequired $false)
 		}
 	}