+++ src/public/Remove-PackageSource.ps1
@@ -1,4 +1,5 @@
 function Remove-PackageSource {
+	[Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSUseShouldProcessForStateChangingFunctions', '', Justification='ShouldProcess support not required by PackageManagement API spec')]
 	param (
 		[Parameter(Mandatory=$true)]
 		[ValidateNotNullOrEmpty()]
@@ -6,9 +7,9 @@ function Remove-PackageSource {
 		$Name
 	)
 
-	Write-Debug ('Remove-PackageSource')
+	Write-Debug ($LocalizedData.ProviderDebugMessage -f ('Remove-PackageSource'))
 
-	[array]$RegisteredPackageSources = Get-PackageSources
+	[array]$RegisteredPackageSources = Cobalt\Get-WinGetSource
 
 	# WinGet.exe will not error if the specified source name isn't already registered, so we will do it here instead.
 	if (-not ($RegisteredPackageSources.Name -eq $Name)) {
@@ -18,6 +19,6 @@ function Remove-PackageSource {
 			-ErrorCategory InvalidArgument
 	}
 
-	# Invoke-WinGet will throw an exception if unregistration fails
-	Invoke-WinGet -SourceRemove -SourceName $Name
+	# Cobalt will throw an exception if unregistration fails
+	Cobalt\Unregister-WinGetSource -Name $Name
 }